Fix a type safety bug.

+++ Sources/Subs.php
@@ -4150,7 +4150,7 @@ function setupMenuContext()
  */
 function smf_seed_generator()
 {
-	updateSettings(array('rand_seed' => microtime() * 1000000));
+	updateSettings(array('rand_seed' => (float) microtime() * 1000000));
 }
 
 /**
